MNM-91-refactor: 마감시간 계산 함수 파일 분리
